Retrofit older cli wrappers
