+++ applications/main/infrared/application.fam
@@ -31,6 +31,10 @@ App(
     apptype=FlipperAppType.STARTUP,
     targets=["f7"],
     entry_point="infrared_on_system_start",
-    sources=["infrared_start.c"],
+    sources=[
+        "infrared_cli.c",
+        "infrared_brute_force.c",
+        "infrared_signal.c",
+    ],
     order=20,
 )
