+++ applications/main/lfrfid/application.fam
@@ -25,6 +25,6 @@ App(
     targets=["f7"],
     apptype=FlipperAppType.STARTUP,
     entry_point="lfrfid_on_system_start",
-    sources=["lfrfid_start.c"],
+    sources=["lfrfid_cli.c"],
     order=50,
 )
