+++ applications/main/lfrfid/lfrfid_cli.c
@@ -14,6 +14,18 @@
 #include <lfrfid/lfrfid_raw_file.h>
 #include <toolbox/pulse_protocols/pulse_glue.h>
 
+static void lfrfid_cli(Cli* cli, FuriString* args, void* context);
+
+#include <cli/cli_i.h>
+CLI_PLUGIN_WRAPPER("lfrfid", lfrfid_cli)
+
+// app cli function
+void lfrfid_on_system_start(void) {
+    Cli* cli = furi_record_open(RECORD_CLI);
+    cli_add_command(cli, "rfid", CliCommandFlagDefault, lfrfid_cli_wrapper, NULL);
+    furi_record_close(RECORD_CLI);
+}
+
 static void lfrfid_cli_print_usage(void) {
     printf("Usage:\r\n");
     printf("rfid read <optional: normal | indala>         - read in ASK/PSK mode\r\n");
@@ -567,17 +579,4 @@ static void lfrfid_cli(Cli* cli, FuriString* args, void* context) {
     }
 
     furi_string_free(cmd);
-}
-
-#include <flipper_application/flipper_application.h>
-#include <cli/cli_i.h>
-
-static const FlipperAppPluginDescriptor plugin_descriptor = {
-    .appid = CLI_PLUGIN_APP_ID,
-    .ep_api_version = CLI_PLUGIN_API_VERSION,
-    .entry_point = &lfrfid_cli,
-};
-
-const FlipperAppPluginDescriptor* lfrfid_cli_plugin_ep(void) {
-    return &plugin_descriptor;
 }
