+++ applications/main/nfc/application.fam
@@ -352,6 +352,6 @@ App(
     targets=["f7"],
     apptype=FlipperAppType.STARTUP,
     entry_point="nfc_on_system_start",
-    sources=["nfc_start.c"],
+    sources=["nfc_cli.c"],
     order=30,
 )
