+++ applications/main/nfc/nfc_cli.c
@@ -63,15 +63,15 @@ static void nfc_cli(Cli* cli, FuriString* args, void* context) {
     furi_string_free(cmd);
 }
 
-#include <flipper_application/flipper_application.h>
 #include <cli/cli_i.h>
+CLI_PLUGIN_WRAPPER("nfc", nfc_cli)
 
-static const FlipperAppPluginDescriptor plugin_descriptor = {
-    .appid = CLI_PLUGIN_APP_ID,
-    .ep_api_version = CLI_PLUGIN_API_VERSION,
-    .entry_point = &nfc_cli,
-};
-
-const FlipperAppPluginDescriptor* nfc_cli_plugin_ep(void) {
-    return &plugin_descriptor;
+void nfc_on_system_start(void) {
+#ifdef SRV_CLI
+    Cli* cli = furi_record_open(RECORD_CLI);
+    cli_add_command(cli, "nfc", CliCommandFlagDefault, nfc_cli_wrapper, NULL);
+    furi_record_close(RECORD_CLI);
+#else
+    UNUSED(nfc_cli);
+#endif
 }
