+++ applications/main/onewire/application.fam
@@ -11,6 +11,5 @@ App(
     appid="onewire_start",
     apptype=FlipperAppType.STARTUP,
     entry_point="onewire_on_system_start",
-    sources=["onewire_start.c"],
     order=60,
 )
