+++ applications/main/subghz/application.fam
@@ -59,7 +59,7 @@ App(
     targets=["f7"],
     apptype=FlipperAppType.STARTUP,
     entry_point="subghz_on_system_start",
-    # sources=["subghz_start.c"],
+    sources=["subghz_cli.c", "helpers/subghz_chat.c"],
     order=40,
 )
