+++ applications/services/crypto/application.fam
@@ -11,6 +11,5 @@ App(
     appid="crypto_start",
     apptype=FlipperAppType.STARTUP,
     entry_point="crypto_on_system_start",
-    sources=["crypto_start.c"],
     order=10,
 )
