+++ applications/services/input/input.c
@@ -80,10 +80,7 @@ const char* input_get_type_name(InputType type) {
 }
 
 #include <cli/cli_i.h>
-
-static void input_cli_wrapper(Cli* cli, FuriString* args, void* context) {
-    cli_plugin_wrapper("input", cli, args, context);
-}
+CLI_PLUGIN_WRAPPER("input", input_cli)
 
 int32_t input_srv(void* p) {
     UNUSED(p);
@@ -102,8 +99,6 @@ int32_t input_srv(void* p) {
 #ifdef SRV_CLI
     Cli* cli = furi_record_open(RECORD_CLI);
     cli_add_command(cli, "input", CliCommandFlagParallelSafe, input_cli_wrapper, event_pubsub);
-#else
-    UNUSED(input_cli_wrapper);
 #endif
 
     InputPinState pin_states[input_pins_count];
