+++ applications/system/js_app/application.fam
@@ -41,8 +41,8 @@ App(
     appid="js_app_start",
     apptype=FlipperAppType.STARTUP,
     entry_point="js_app_on_system_start",
-    sources=["js_start.c"],
     order=160,
+    sources=["js_app.c"],
 )
 
 App(
